Moved components to new file
